+++ extensions/quick-access/src/utils/constants.ts
@@ -30,3 +30,8 @@ export const imgExt = [
   ".tiff",
   ".webp",
 ];
+
+export const ZIP_EXT = [".zip", ".rar", ".7z", ".tar", ".gz", ".bz2", ".xz", ".iso"];
+export const NO_PREVIEW_EXTENSIONS = [".exe", ".DS_Store", ".gdoc", ".gsheet", ".gslides", ".ts", ".css"];
+export const DMG_EXT = ".dmg";
+export const APP_EXT = ".app";