+++ extensions/capture-raycast-metadata/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Capture Raycast Metadata Changelog
 
-## [Added Raycast Wallpaper] - 2022-04-25
+## [Update Capture Raycast Metadata] - 2022-05-04
+
+- Increase the delay of Toast pop-ups to 0.3 second.
+
+## [Update Capture Raycast Metadata] - 2022-04-25
 
 - Add preference Screenshots Directory, you can set the directory where the screenshots are saved
 
-## [Added Raycast Wallpaper] - 2022-04-15
+## [Added Capture Raycast Metadata] - 2022-04-15
 
 - Release of Capture Raycast Metadata version 1.0
 - **Capture Raycast Metadata**: Capture a screenshot for Raycast Extension's metadata.