+++ extensions/todoist/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - Added the ability to add a new task from a project
 - Shared the task actions between the task detail and the task list item
+## [Edit your projects] - 2022-04-30
+
+- Added a new action on projects to edit them
+- Added a new action on projects to add/remove them to/from favorites
+- Added the project's color on the Search Projects command
+- Removed the custom ordering of the tasks since Todoist already performs them
+- Displayed the sub-tasks back in the project view
 
 ## [Add support for comments] - 2022-04-23