+++ extensions/visual-studio-code-recent-projects/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Visual Studio Code Search Recent Projects Changelog
 
+## [Updates] - 2022-05-12
+
+Fixed missing workspaces from previous update.
+
+## [Updates] - 2022-05-09
+
+Adds sqllite db support for recent files/folders which is used from Visual Studio Code v1.65.0.
+
 ## [Updates] - 2021-10-13
 
 Fixed show in finder action, and added development actions.